Add files via upload
This file looks at data for runs 133-137. It plots a histogram of intensity around the first water ring on a shot-to-hot basis. background threshold are set to Intensity values of less than the median and hit thresholds are set to an intensity value of twice the background threshold. It also outputs tag number of the hits and saves the OFF shots average